pagination for admin data list and registry.go refactored
