+++ gojang/admin/views/css/admin.css
@@ -131,15 +131,142 @@
 /* Tables */
 .admin-table { width: 100%; border-collapse: collapse; }
 .admin-table thead { background: #f8fafc; border-bottom: 2px solid #e2e8f0; }
-.admin-table th { padding: 0.75rem 1rem; text-align: left; font-weight: 600; color: #475569; font-size: 0.875rem; text-transform: uppercase; letter-spacing: 0.05em; }
+.admin-table th { padding: 0.5rem 0.75rem; text-align: left; font-weight: 600; color: #475569; font-size: 0.8125rem; text-transform: uppercase; letter-spacing: 0.05em; }
 .admin-table tbody tr { border-bottom: 1px solid #e2e8f0; transition: background 0.15s; }
 .admin-table tbody tr:hover { background: #f8fafc; }
-.admin-table td { padding: 1rem; color: #334155; }
+.admin-table td { padding: 0.5rem 0.75rem; color: #334155; font-size: 0.875rem; }
 .admin-actions-col { width: 150px; text-align: right; }
 .admin-action-buttons { display: flex; gap: 0.5rem; justify-content: flex-end; }
 .admin-empty-state { text-align: center; padding: 3rem 1rem !important; color: #64748b; }
 .admin-empty-state a { color: #3b82f6; text-decoration: none; font-weight: 500; }
 .admin-empty-state a:hover { text-decoration: underline; }
+
+/* Table controls and pagination */
+.admin-table-controls {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 1rem;
+    background: #f8fafc;
+    border-bottom: 1px solid #e2e8f0;
+}
+
+.admin-controls-left {
+    display: flex;
+    align-items: center;
+}
+
+.admin-count-label {
+    color: #64748b;
+    font-size: 0.9rem;
+    font-weight: 500;
+}
+
+.admin-controls-right {
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+}
+
+.admin-per-page-label {
+    color: #64748b;
+    font-size: 0.9rem;
+    font-weight: 500;
+}
+
+.admin-select {
+    padding: 0.5rem 2rem 0.5rem 0.75rem;
+    border: 1px solid #cbd5e1;
+    border-radius: 0.375rem;
+    background: white;
+    color: #334155;
+    font-size: 0.875rem;
+    cursor: pointer;
+    transition: all 0.15s;
+    appearance: none;
+    background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%236b7280' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
+    background-position: right 0.5rem center;
+    background-repeat: no-repeat;
+    background-size: 1.5em 1.5em;
+}
+
+.admin-select:hover {
+    border-color: #94a3b8;
+}
+
+.admin-select:focus {
+    outline: none;
+    border-color: #3b82f6;
+    box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.1);
+}
+
+.admin-pagination {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 1rem;
+    background: #f8fafc;
+    border-top: 1px solid #e2e8f0;
+}
+
+.admin-page-info {
+    color: #64748b;
+    font-size: 0.9rem;
+    font-weight: 500;
+}
+
+.admin-page-buttons {
+    display: flex;
+    gap: 0.25rem;
+    align-items: center;
+}
+
+.admin-btn-page {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    min-width: 2.5rem;
+    height: 2.5rem;
+    padding: 0.5rem 0.75rem;
+    border: 1px solid #cbd5e1;
+    border-radius: 0.375rem;
+    background: white;
+    color: #475569;
+    font-size: 0.875rem;
+    font-weight: 500;
+    text-decoration: none;
+    transition: all 0.15s;
+    cursor: pointer;
+}
+
+.admin-btn-page:hover {
+    background: #f1f5f9;
+    border-color: #94a3b8;
+    color: #1e293b;
+}
+
+.admin-btn-page.active {
+    background: #3b82f6;
+    border-color: #3b82f6;
+    color: white;
+    cursor: default;
+}
+
+.admin-btn-page.active:hover {
+    background: #3b82f6;
+    border-color: #3b82f6;
+}
+
+.admin-page-ellipsis {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    min-width: 2.5rem;
+    height: 2.5rem;
+    color: #94a3b8;
+    font-weight: 500;
+}
+
 
 /* Modals & forms (admin-specific) */
 .admin-modal-overlay, .admin-form-modal-overlay { position: fixed; top: 0; left: 0; right: 0; bottom: 0; background: rgba(0, 0, 0, 0.5); display: flex; align-items: center; justify-content: center; z-index: 1000; animation: fadeIn 0.2s; }
